# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Beasiswa API Dokumentasi",
        "description": "Dokumentasi API untuk Sistem Beasiswa Universitas Esa Unggul",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/beasiswa": {
            "get": {
                "tags": [
                    "Beasiswa"
                ],
                "summary": "Menampilkan semua program beasiswa",
                "operationId": "7669e0f0447ba7dd8aed4f4b7d930c5f",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data beasiswa"
                    }
                }
            },
            "post": {
                "tags": [
                    "Beasiswa"
                ],
                "summary": "Menyimpan program beasiswa baru",
                "operationId": "3ced3b084183102d2779b54109446400",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_beasiswa",
                                    "periode_buka",
                                    "periode_tutup"
                                ],
                                "properties": {
                                    "nama_beasiswa": {
                                        "type": "string",
                                        "example": "Beasiswa Unggulan"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Beasiswa untuk mahasiswa berprestasi"
                                    },
                                    "periode_buka": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-01"
                                    },
                                    "periode_tutup": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-01"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "aktif"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Beasiswa berhasil dibuat"
                    }
                }
            }
        },
        "/api/beasiswa/{id}": {
            "get": {
                "tags": [
                    "Beasiswa"
                ],
                "summary": "Menampilkan detail program beasiswa",
                "operationId": "c13a670c7afc49b90e27b1894c55208e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail beasiswa"
                    }
                }
            },
            "put": {
                "tags": [
                    "Beasiswa"
                ],
                "summary": "Memperbarui program beasiswa",
                "operationId": "3f10cdaae91a6b01522d6ba042fabe4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_beasiswa": {
                                        "type": "string",
                                        "example": "Beasiswa Update"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Deskripsi terbaru"
                                    },
                                    "periode_buka": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-10"
                                    },
                                    "periode_tutup": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-10"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "nonaktif"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Beasiswa berhasil diperbarui"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Beasiswa"
                ],
                "summary": "Menghapus program beasiswa",
                "operationId": "6890f7db211a2787b1e9985baabadde8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Beasiswa berhasil dihapus"
                    }
                }
            }
        },
        "/api/data-peserta": {
            "get": {
                "tags": [
                    "Data Peserta"
                ],
                "summary": "Menampilkan semua data peserta",
                "operationId": "8b1ee8ec34fb3a5c2f533f89d15915af",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data peserta"
                    }
                }
            },
            "post": {
                "tags": [
                    "Data Peserta"
                ],
                "summary": "Menyimpan data peserta baru",
                "operationId": "54ccb6491a97dd3bcf13e9898919f1be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "nim",
                                    "jurusan",
                                    "angkatan"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nim": {
                                        "type": "string",
                                        "example": "20230801219"
                                    },
                                    "jurusan": {
                                        "type": "string",
                                        "example": "Teknik Informatika"
                                    },
                                    "angkatan": {
                                        "type": "integer",
                                        "example": 2023
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data peserta berhasil dibuat"
                    }
                }
            }
        },
        "/api/data-peserta/{id}": {
            "get": {
                "tags": [
                    "Data Peserta"
                ],
                "summary": "Menampilkan detail data peserta",
                "operationId": "c56e240c33a4379e3e72f0498c02c3c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail data peserta"
                    }
                }
            },
            "put": {
                "tags": [
                    "Data Peserta"
                ],
                "summary": "Memperbarui data peserta",
                "operationId": "dd897e797780a1a32afe2b8fcfa509ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nim": {
                                        "type": "string",
                                        "example": "20230801219"
                                    },
                                    "jurusan": {
                                        "type": "string",
                                        "example": "Sistem Informasi"
                                    },
                                    "angkatan": {
                                        "type": "integer",
                                        "example": 2022
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data peserta berhasil diperbarui"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Data Peserta"
                ],
                "summary": "Menghapus data peserta",
                "operationId": "578e2e70e030230f4f08bfb57b275065",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Data peserta berhasil dihapus"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}